fix: rename inputs components
